refactor: reorganize components into feature-based directory structure
feat(invoice): add new invoice-related components and composables
feat(items): add variants component for item selection
feat(payment): add mpesa payments component
feat(customer): add new address component
feat(dialogs): add drafts dialog component
feat(mixins): add invoice shortcuts mixin

fix: update import paths to reflect new structure
fix: remove duplicate components from old pos directory
fix: clean up unused files and consolidate logic

chore: move existing components to new organized structure
