+++ posawesome/public/js/posapp/components/customer/Customer.vue
