+++ posawesome/public/js/posapp/components/dialogs/ClosingDialog.vue
