+++ posawesome/public/js/posapp/components/dialogs/OpeningDialog.vue
