+++ posawesome/public/js/posapp/components/items/CameraScanner.vue
