+++ posawesome/public/js/posapp/components/items/ItemsSelector.vue
