+++ posawesome/public/js/posapp/components/items/Variants.vue
