+++ posawesome/public/js/posapp/components/offers/PosOffers.vue
