--- posawesome/public/js/posapp/components/pos/invoiceShortcuts.js
@@ -1,90 +1,0 @@
-export default {
-    shortOpenFirstItem(e) {
-      if (e.key.toLowerCase() === "a" && (e.ctrlKey || e.metaKey)) {
-        try {
-          e.preventDefault();
-          e.stopPropagation();
-
-          if (!this.items || this.items.length === 0) {
-            console.log('No items to expand/collapse');
-            return;
-          }
-
-          const firstItem = this.items[0];
-          console.log('Processing first item:', firstItem.item_code);
-
-          // Check if first item is currently expanded using its ID
-          const isExpanded = this.expanded.includes(firstItem.posa_row_id);
-
-          // Toggle expanded state using item ID
-          if (isExpanded) {
-            console.log('Collapsing item:', firstItem.item_code);
-            this.expanded = [];
-          } else {
-            console.log('Expanding item:', firstItem.item_code);
-            this.expanded = [firstItem.posa_row_id];
-            // Update item details when expanding
-            this.$nextTick(() => {
-              this.update_item_detail(firstItem);
-            });
-          }
-        } catch (error) {
-          console.error('Error in shortOpenFirstItem:', error);
-          this.eventBus.emit("show_message", {
-            title: __("Error toggling item details"),
-            color: "error"
-          });
-        }
-      }
-    },
-
-    handleExpandedUpdate(newExpanded) {
-      console.log('Expanded state updated:', newExpanded);
-      this.expanded = newExpanded;
-
-      // Update item details for newly expanded items
-      if (newExpanded && newExpanded.length > 0) {
-        const expandedItemId = newExpanded[0];
-        const expandedItem = this.items.find(item => item.posa_row_id === expandedItemId);
-        if (expandedItem) {
-          this.$nextTick(() => {
-            this.update_item_detail(expandedItem);
-          });
-        }
-      }
-    },
-
-
-
-    // Keyboard shortcut: open payment dialog
-    shortOpenPayment(e) {
-      if (e.key === "s" && (e.ctrlKey || e.metaKey)) {
-        e.preventDefault();
-        this.show_payment();
-      }
-    },
-
-    // Keyboard shortcut: delete first item from the invoice
-    shortDeleteFirstItem(e) {
-      if (e.key === "d" && (e.ctrlKey || e.metaKey)) {
-        e.preventDefault();
-        this.remove_item(this.items[0]);
-      }
-    },
-
-    shortSelectDiscount(e) {
-      console.log('Shortcut pressed:', e.key, e.ctrlKey);
-      if (e.key.toLowerCase() === "e" && (e.ctrlKey || e.metaKey)) {
-        console.log('Focusing discount field');
-        e.preventDefault();
-        e.stopPropagation();
-        if (this.$refs.discount) {
-          this.$refs.discount.focus();
-          console.log('Discount field focused');
-        } else {
-          console.log('Discount field ref not found');
-        }
-      }
-    },
-
-};
